+++ src/module/entities/TwodsixActor.ts
@@ -300,8 +300,9 @@ export default class TwodsixActor extends Actor {
       case 'space-object':
         break;
       default:
-        ui.notifications.error(game.i18n.localize("Twodsix.Actor.UnknownActorType") + " " + this.type);
-        console.error(`Unknown actor type: ${this.type}`);
+        //Don't push error as it breaks data prep.  Needed for other modules that create types
+        //ui.notifications.error(game.i18n.localize("Twodsix.Actor.UnknownActorType") + " " + this.type);
+        console.warn(`Unknown actor type: ${this.type}`);
     }
   }